feat: sync product page with cart
