+++ src/pages/ProductPage/ProductPage.tsx
@@ -1,6 +1,4 @@
-import { useCallback, useEffect, useState } from 'react';
-
-import { useSelector } from 'react-redux';
+import { useEffect, useState } from 'react';
 
 import getAttribute from './lib/helpers/getAttribute.ts';
 import AddWishlistMobile from './ui/AddWishlistMobile.tsx';
@@ -17,9 +15,6 @@ import Price from './ui/Price.tsx';
 import Rating from './ui/Rating.tsx';
 import Title from './ui/Title.tsx';
 import TitleAbout from './ui/TitleAbout.tsx';
-import { RootState } from '../../app/store';
-import { useAddLineItemMutation, useLazyGetCartByIdQuery } from '../../entities/cart';
-import { AddLineItemRequestBody, RemoveLineItemRequestBody } from '../../entities/cart/types/types.ts';
 import { ProductAttributeNames, useGetProductQuery } from '../../entities/product';
 import 'swiper/css';
 import { ProductPrice } from '../../entities/product/types/types.ts';
@@ -32,66 +27,6 @@ export default function ProductPage() {
   const [isSliderOpen, setSliderOpen] = useState(false);
   const productId = useGetPath();
   const { data } = useGetProductQuery(productId);
-  const cartId = useSelector((state: RootState) => state.userReducer.cartId);
-  const [getCart, { data: cart }] = useLazyGetCartByIdQuery();
-  const [updateLineItem, { data: newCart }] = useAddLineItemMutation();
-
-  const memoizedGetCart = useCallback(
-    (_cartId: string) => {
-      getCart(_cartId, false);
-    },
-    [getCart],
-  );
-
-  useEffect(() => {
-    memoizedGetCart(cartId);
-  }, [cartId, memoizedGetCart, newCart]);
-
-  const addToCart = async () => {
-    const body: AddLineItemRequestBody = {
-      version: cart?.version || 1,
-      actions: [
-        {
-          action: 'addLineItem',
-          productId,
-          variantId: 1,
-          quantity: 1,
-        },
-      ],
-    };
-    try {
-      const result = await updateLineItem({ cartId, body }).unwrap();
-      return result;
-    } catch (e) {
-      // throw new Error(e);
-    }
-    return null;
-  };
-
-  const removeOneFromCart = async () => {
-    try {
-      const targetItem = cart?.lineItems.find((item) => item.productId === productId);
-      if (targetItem) {
-        const { lineItemId } = targetItem;
-        const body: RemoveLineItemRequestBody = {
-          version: cart?.version || 1,
-          actions: [
-            {
-              action: 'removeLineItem',
-              lineItemId,
-              variantId: 1,
-              quantity: 1,
-            },
-          ],
-        };
-        const result = await updateLineItem({ cartId, body }).unwrap();
-        return result;
-      }
-    } catch (e) {
-      // throw new Error(e);
-    }
-    return null;
-  };
 
   /// add to cart btn will change to remove from cart an this function will be called
   // const removeAllFromCart = async () => {
@@ -192,7 +127,7 @@ export default function ProductPage() {
                   <Price centOldPrice={centOldPrice} centPrice={centPrice} currencyCode={currencyCode} />
                 </>
               </Header>
-              <Footer addToCart={addToCart} removeOneFromCart={removeOneFromCart} />
+              <Footer />
               <Description attributes={attributes} />
               {ingredients ? (
                 <IngredientList>
