+++ src/shared/lib/helpers/index.ts
@@ -2,6 +2,7 @@ import capitalize from './capitalize.ts';
 import deleteCookie from './deleteCookie.ts';
 import getCookie from './getCookie.ts';
 import getErrorMessage from './getErrorMessage.ts';
+import padZero from './padZero.ts';
 import setCookie from './setCookie.ts';
 
-export { setCookie, getCookie, deleteCookie, getErrorMessage, capitalize };
+export { setCookie, getCookie, deleteCookie, getErrorMessage, capitalize, padZero };
